Log in and sign up system inside a meteors game with high scores board
In this project there is a game of a spaceship that escapes from meteors, the goal of the user is to reach the finish line at the end of the screen. When the user enters the game he must register, after registration his data will be saved in a csv file. He will play the game and get a score (how long he played the game). After reaching the finish line, a name board of the five players with the best score will appear.  The libraries are consumed: pandas, pygame and pyautogui.